feat: implement AttachDiaryFileFromContentService
