+++ src/main/kotlin/me/daegyeo/maru/diary/application/port/in/AttachDiaryFileFromContentUseCase.kt
@@ -1,0 +1,7 @@
+package me.daegyeo.maru.diary.application.port.`in`
+
+import me.daegyeo.maru.diary.application.port.`in`.command.AttachDiaryFileFromContentCommand
+
+fun interface AttachDiaryFileFromContentUseCase {
+    fun attachDiaryFileFromContent(input: AttachDiaryFileFromContentCommand)
+}
